requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
geographiclib==2.0
geopy==2.3.0
inflection==0.5.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.6.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
rpds-py==0.8.8
sqlparse==0.4.3
uritemplate==4.1.1
